requirements.txt
# Gateway Service Dependencies - Pinned for Production
Django>=5.1,<5.3
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-oauth-toolkit>=3.0,<4.0
django-cors-headers>=4.4,<5.0
django-filter>=24.0,<26.0
django-redis>=5.4,<7.0
django-celery-results>=2.5,<3.0
django-celery-beat>=2.6,<3.0
drf-spectacular>=0.27,<1.0
whitenoise>=6.7,<7.0
psycopg2-binary>=2.9,<3.0
celery>=5.4,<6.0
redis>=5.0,<8.0
requests>=2.32,<3.0
gunicorn>=22.0,<26.0
python-dotenv>=1.0,<2.0

# Testing
pytest>=8.0
pytest-django>=4.8
pytest-cov>=5.0
